feat: switched app context to only do one thing && file name changes
