+++ src/UI/components/common/File.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Flex, HStack, Menu, Text, VStack } from '@chakra-ui/react';
 import { TbFolder, TbFile } from 'react-icons/tb';
-import { useSelectedFSNodeFile } from '../../context/selected-fs-node-context';
+import { useSelectedFSNodeFile } from '../../context/selected.fs.node.context';
 import ContextMenu from '../compound/ContextMenu';
 import { useContextMenu } from '../../hooks/useContextMenu';
 import { FileSystemNode } from '../../../core/entities/file.system.node.entity';
