+++ src/UI/components/common/Folder.tsx
@@ -3,8 +3,8 @@ import { Flex, HStack, Text, VStack } from '@chakra-ui/react';
 import { TbFolder } from 'react-icons/tb';
 import { formatDate } from '../../../utils/date';
 import { FileSystemNode } from '../../../core/entities/file.system.node.entity';
-import { useTargetFolder } from '../../context/target-folder';
-import { useSelectedFSNodeFile } from '../../context/selected-fs-node-context';
+import { useTargetFolder } from '../../context/target.folder.context';
+import { useSelectedFSNodeFile } from '../../context/selected.fs.node.context';
 
 type Props = {
   file: FileSystemNode;
