+++ src/UI/components/compound/SideNav.tsx
@@ -6,7 +6,7 @@ import MenuItem from '../../interfaces/Menu';
 import SideMenuItem from '../common/SideMenuItem';
 import { Link } from 'react-router-dom';
 import StorageStats from './StorageStats';
-import { useSelectedFSNodeFile } from '../../context/selected-fs-node-context';
+import { useSelectedFSNodeFile } from '../../context/selected.fs.node.context';
 
 const SideNav = () => {
   const [name, setName] = React.useState('');
