--- src/UI/context/index.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import { AuthProvider } from './auth-context';
-import { UserProvider } from './user-context';
-
-type Props = {
-  children: React.ReactNode;
-};
-
-const AppProviders = ({ children }: Props) => {
-  return (
-    <AuthProvider>
-      <UserProvider>{children}</UserProvider>
-    </AuthProvider>
-  );
-};
-
-export default AppProviders;
