+++ src/index.tsx
@@ -1,13 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-import { AuthProvider } from './UI/context/auth-context';
-import { UserProvider } from './UI/context/user-context';
-import { SelectedFSNodeProvider } from './UI/context/selected-fs-node-context';
-import './index.css';
-import { TargetFolderProvider } from './UI/context/target-folder';
+import AppProvider from '@context/app.context';
+import App from '@src/App';
+import '@src/index.css';
 
 const queryClient = new QueryClient();
 
@@ -17,20 +13,9 @@ const root = ReactDOM.createRoot(
 root.render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
-      <AuthProvider>
-        <UserProvider>
-          <SelectedFSNodeProvider>
-            <TargetFolderProvider>
-              <App />
-            </TargetFolderProvider>
-          </SelectedFSNodeProvider>
-        </UserProvider>
-      </AuthProvider>
+      <AppProvider>
+        <App />
+      </AppProvider>
     </QueryClientProvider>
   </React.StrictMode>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
