+++ tsconfig.json
@@ -13,6 +13,9 @@
       "@dto/*": ["src/core/dto/*"],
       "@services/*": ["src/core/services/*"],
       "@useCases/*": ["src/useCases/*"],
+      "@context/*": ["src/UI/context/*"],
+      "@hooks/*": ["src/UI/hooks/*"],
+      "@config/*": ["src/config/*"],
     },
     "allowJs": true,
     "skipLibCheck": true,
